更新 1.系統訊息 2.輸入檢查 3.README.md

+++ line_bot/commands.py
@@ -13,13 +13,15 @@ line_bot_api = LineBotApi(env('channel_access_token'))
 handler = WebhookHandler(env('channel_secret'))
 
 
-def reply_line(event, content):
-  return line_bot_api.reply_message(
-    event.reply_token,
-    TextSendMessage(text=content)
-)
+SITE = "https://github.com/BWsix/line-to-discord"
+
 
 class Commands:
+  
+  def reply_line(self, event, content):
+    return line_bot_api.reply_message(
+      event.reply_token,
+      TextSendMessage(text=content))
   
   def get_group(self, event) -> Group:
     try:
@@ -34,9 +36,9 @@ class Commands:
     group = self.get_group(event)
 
     if group is None:
-      return reply_line(event, "這個群組尚未啟用discohook!\n使用指令 'discohook' 啟用")
+      return self.reply_line(event, "這個群組尚未啟用discohook!\n使用指令 'discohook' 啟用")
     if event.source.user_id != group.manager:
-      return reply_line(event, "只有管理員能夠使用指令!")
+      return self.reply_line(event, "只有管理員能夠使用指令!")
     
     return group
 
@@ -44,7 +46,7 @@ class Commands:
   def handle_create(self, event):
     try:
       Group.objects.get(id=event.source.group_id)
-      return reply_line(event, "指令無效 : discohook已經啟用")
+      return self.reply_line(event, "指令無效 : discohook已經啟用")
     except Group.DoesNotExist:
       pass
 
@@ -60,30 +62,69 @@ class Commands:
       manager=event.source.user_id,
     )
 
-    return reply_line(event, "discohook啟用成功\n將我加入好友便能在discord顯示你的姓名")
+    return self.reply_line(event, f"啟用成功\n將本bot加入好友便能在discord顯示你的姓名\n請到以下網址查看指令列表\n{SITE}")
   
   def handle_link(self, event):
     group = self.admin_checker(event)
+    texts = event.message.text.split('\n')
+    
+    if len(texts) != 3:
+      return self.reply_line(event, f"指令格式有誤，請到以下網址查看指令列表\n{SITE}")
+    if len(texts[1]) > 20:
+      return self.reply_line(event, f"名稱太長\n必須小於20字")
+    
+    try:
+      hook = Hook.objects.get(name=texts[1])
+      return self.reply_line(event, f"這個名稱已被註冊")
+    except Hook.DoesNotExist:
+      pass
+
+    try:
+      hook = Hook.objects.get(url=texts[2])
+      return self.reply_line(event, f"這個頻道已被{hook.name}註冊")
+    except Hook.DoesNotExist:
+      pass
 
     Hook.objects.create(
       owner=group,
-      name=event.message.text.split('\n')[1],
-      url=event.message.text.split('\n')[2],
+      name=texts[1],
+      url=texts[2],
     )
 
-    return reply_line(event, "加入完成\n現在開始所有傳送至此群組的訊息將會藉由剛才提供的連結一併傳送至discord群組")
+    try:
+      summary = line_bot_api.get_group_summary(event.source.group_id)
+      group_name = summary.group_name
+    except LineBotApiError:
+      group_name = "unknown"
+    post(texts[2], data={'content':f"(**{group_name}已建立連結**)"})
+    
+    return self.reply_line(event, "加入完成\n現在開始所有傳送至此群組的訊息將會藉由剛才提供的連結一併傳送至discord群組")
   
   def handle_unlink(self, event):
-    group = self.admin_checker(event)
-    name = event.message.text.split('\n')[1]
+    self.admin_checker(event)
+    texts = event.message.text.split('\n')
+    
+    if len(texts) != 2:
+      return self.reply_line(event, f"指令格式有誤，請到以下網址查看指令列表\n{SITE}")
+    
+    name = texts[1]
     
     try:
       object = Hook.objects.get(name=name)
+
+      try:
+        summary = line_bot_api.get_group_summary(event.source.group_id)
+        group_name = summary.group_name
+      except LineBotApiError:
+        group_name = "unknown"
+      
+      post(object.url, data={'content':f"(**{group_name}已解除連結**)"})
       object.delete()
-    except Group.DoesNotExist:
-      return reply_line(event, f"{name}不在discohook名單中\n使用指令 'discoquery' 取得discohook名單")
+    except Hook.DoesNotExist:
+      return self.reply_line(event, f"{name}不在已連結群組名單中\n使用指令 'discoquery' 取得已連結群組名單")
     
-    return reply_line(event, f"成功解除連結\n將不會傳送訊至{name}")
+
+    return self.reply_line(event, f"成功解除連結\n將不會傳送訊至{name}")
     
   def handle_query(self, event):
     group = self.admin_checker(event)
@@ -93,7 +134,7 @@ class Commands:
     for item in hooks:
       hook_list += item.name + '\n'
     
-    return reply_line(event, hook_list)
+    return self.reply_line(event, hook_list)
 
 
   def post(self, event, **kwargs):
